+++ src/cssHelper.js
@@ -32,6 +32,12 @@ class RecrasCSSHelper {
 .booking-error, .minimum-amount {
     color: hsl(0, 50%, 50%);
 }
+.booking-error {
+    -ms-grid-column: 1;
+    -ms-grid-column-span: 2;
+    grid-column: 1 / 3;
+    -ms-grid-row: 999; /* No specified row will place it on top of certain elements. This will put it at the very end */
+}
 .minimum-amount {
     padding-left: 0.5em;
 }